198 - fix for logged in customer with empty cart + translates

+++ Controller/Adminhtml/Customer/SaveCompanyId.php
@@ -10,7 +10,6 @@ declare(strict_types=1);
 namespace Hokodo\BNPL\Controller\Adminhtml\Customer;
 
 use Hokodo\BNPL\Api\Data\OrganisationInterface;
-use Hokodo\BNPL\Api\Data\UserInterface;
 use Hokodo\BNPL\Api\HokodoCustomerRepositoryInterface;
 use Hokodo\BNPL\Api\HokodoQuoteRepositoryInterface;
 use Hokodo\BNPL\Gateway\Service\Organisation;
@@ -24,6 +23,7 @@ use Magento\Customer\Api\SessionCleanerInterface;
 use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\Controller\Result\JsonFactory;
 use Magento\Framework\Controller\ResultInterface;
+use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Quote\Api\CartRepositoryInterface;
 use Psr\Log\LoggerInterface;
 
@@ -133,56 +133,50 @@ class SaveCompanyId extends Action implements HttpPostActionInterface
         $hokodoCustomer = $this->hokodoCustomerRepository->getByCustomerId($customerId);
         $hokodoCustomer->setCompanyId($companyId);
 
-        if (!$hokodoCustomer || !$hokodoCustomer->getId()) {
-            $hokodoCustomer
-                ->setCustomerId($customerId)
-                ->setOrganisationId('')
-                ->setUserId('');
-        }
-
         try {
             /* @todo remove this and similar piec of code from webapi into on service */
-            if (!$hokodoCustomer->getOrganisationId()) {
-                /** @var CustomerRepositoryInterface $customer */
-                $customer = $this->customerRepository->getById($customerId);
-                /** @var OrganisationInterface $organisation */
-                $organisation = $this->organisationService->createOrganisation(
-                    $this->organisationBuilder->build(
-                        $companyId,
-                        $customer->getEmail()
-                    )
-                )->getDataModel();
-                $hokodoCustomer->setOrganisationId($organisation->getId());
-            }
+            /** @var CustomerRepositoryInterface $customer */
+            $customer = $this->customerRepository->getById($customerId);
+            /** @var OrganisationInterface $organisation */
+            $organisation = $this->organisationService->createOrganisation(
+                $this->organisationBuilder->build(
+                    $companyId,
+                    $customer->getEmail()
+                )
+            )->getDataModel();
+            $hokodoCustomer->setOrganisationId($organisation->getId());
 
-            if (!$hokodoCustomer->getUserId()) {
-                if (empty($customer)) {
-                    $customer = $this->customerRepository->getById($customerId);
-                }
-                /** @var UserInterface $user */
-                $user = $this->userService->createUser(
-                    $this->userBuilder->build($customer, $organisation->getId())
-                )->getDataModel();
-                $hokodoCustomer->setUserId($user->getId());
-            }
+            $user = $this->userService->createUser(
+                $this->userBuilder->build($customer, $organisation->getId())
+            )->getDataModel();
+            $hokodoCustomer->setUserId($user->getId());
+
             $this->hokodoCustomerRepository->save($hokodoCustomer);
+
+            /* @var \Magento\Quote\Api\Data\CartInterface $cart */
+            try {
+                $cart = $this->cartRepository->getActiveForCustomer($customerId);
+                if ($cart->getId()) {
+                    $hokodoQuote = $this->hokodoQuoteRepository->getByQuoteId($cart->getId());
+                    if ($hokodoQuote->getQuoteId()) {
+                        $this->hokodoQuoteRepository->deleteByQuoteId($cart->getId());
+                    }
+                }
+            } catch (NoSuchEntityException $exception) {
+                $this->logger->notice(
+                    __('Hokodo_BNPL: try get active quote. ', $exception->getMessage())
+                );
+            }
+            $this->sessionCleanerInterface->clearFor($customerId);
             $result = [
                 'success' => true,
                 'message' => __('Company Was Updated.'),
             ];
-
-            /** @var \Magento\Quote\Api\Data\CartInterface $cart */
-            $cart = $this->cartRepository->getActiveForCustomer($customerId);
-            if ($cart->getId()) {
-                $this->hokodoQuoteRepository->deleteByQuoteId($cart->getId());
-            }
-            $this->sessionCleanerInterface->clearFor($customerId);
-            return $resultJson->setData($result);
         } catch (\Exception $e) {
             $this->logger->critical(
                 __('Hokodo_BNPL: set company to user failed with error - %1', $e->getMessage())
             );
-            return $resultJson->setData($result);
         }
+        return $resultJson->setData($result);
     }
 }
